+++ pywa/types/message_status.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from .callback import _CallbackDataT
+from .user import User as _User
 
 __all__ = [
     "Conversation",
@@ -29,6 +30,36 @@ if TYPE_CHECKING:
     from ..client import WhatsApp
 
 _logger = logging.getLogger(__name__)
+
+
+@dataclasses.dataclass(frozen=True, slots=True)
+class User(_User):
+    """
+    Represents the WhatsApp user a :class:`MessageStatus` was sent to.
+
+    Attributes:
+        bsuid: The WhatsApp user’s BSUID. See `developers.facebook.com <https://developers.facebook.com/documentation/business-messaging/whatsapp/business-scoped-user-ids>`_ for more information. Will be ``None`` if the message status is ``FAILED`` because the user has no WhatsApp account.
+        wa_id: The user's phone number in international format (without the '+' sign). Will be unavailable if the user enables the username feature. See `developers.facebook.com <https://developers.facebook.com/documentation/business-messaging/whatsapp/business-scoped-user-ids#phone-numbers>`_ for more information.
+        name: The name of the user.
+        username: The username of the user.
+        identity_key_hash: The identity key hash of the user (Only if identity key check is enabled on the phone number settings).
+        parent_bsuid: The Parent business-scoped user ID. See `developers.facebook.com <https://developers.facebook.com/documentation/business-messaging/whatsapp/business-scoped-user-ids#parent-business-scoped-user-ids>`_ for more information.
+    """
+
+    bsuid: str | None
+
+    @classmethod
+    def from_contact(cls, data: dict, client: WhatsApp) -> User:
+        profile = data.get("profile", {})
+        return cls(
+            _client=client,
+            bsuid=data.get("user_id"),
+            wa_id=data.get("wa_id") or None,  # avoid empty string
+            name=profile.get("name"),
+            username=profile.get("username"),
+            parent_bsuid=data.get("parent_user_id"),
+            identity_key_hash=data.get("identity_key_hash"),
+        )
 
 
 class MessageStatusType(helpers.StrEnum):
@@ -129,7 +160,7 @@ class MessageStatus(BaseUserUpdate, Generic[_CallbackDataT]):
         metadata: The metadata of the message (to which phone number it was sent).
         status: The status of the message.
         timestamp: The timestamp when the status was updated (in UTC).
-        from_user: The user who the message was sent to.
+        from_user: The user who the message was sent to. The user may not have a WhatsApp account, in which case the ``bsuid`` attribute will be ``None``.
         conversation: The conversation that the message was sent in (See `Conversation <https://developers.facebook.com/docs/whatsapp/pricing#conversations>`_).
         pricing: The pricing of the message (Optional).
         error: The error that occurred (if status is :class:`MessageStatusType.FAILED`).
@@ -137,6 +168,7 @@ class MessageStatus(BaseUserUpdate, Generic[_CallbackDataT]):
         shared_data: Shared data between handlers.
     """
 
+    from_user: User
     status: MessageStatusType
     pricing: Pricing | None
     conversation: Conversation | None
@@ -146,6 +178,7 @@ class MessageStatus(BaseUserUpdate, Generic[_CallbackDataT]):
     _txt_fields = ("tracker",)
     _webhook_field = "messages"
     _is_user_action = False
+    _usr_cls = User
 
     @classmethod
     def from_update(
@@ -167,7 +200,7 @@ class MessageStatus(BaseUserUpdate, Generic[_CallbackDataT]):
             metadata=Metadata.from_dict(value["metadata"]),
             status=MessageStatusType(status["status"]),
             timestamp=helpers.timestamp_to_datetime(status["timestamp"]),
-            from_user=client._usr_cls.from_contact(
+            from_user=cls._usr_cls.from_contact(
                 value["contacts"][contact_idx], client=client
             ),
             tracker=status.get("biz_opaque_callback_data"),